created packages and refactored files within them. Also, refactored the names of some files to make them more succinct to their responsibilities
