+++ gradle.properties
@@ -5,9 +5,9 @@
 group=replace.me
 version=1.0.0-SNAPSHOT
 
-# TODO: Replace "replace.me.project.Main" with the fully qualified class name of the project's main
+# TODO: Replace "replace.me.project.com.islandescape.client.Main" with the fully qualified class name of the project's main
 #  class, without any quotes, as the value of the mainClass property.
-mainClass=Main
+mainClass=com.islandescape.client.com.islandescape.client.Main
 
 javaVersion=11
 shadowVersion=7.1.2
