+++ src/main/java/com/islandescape/controllers/GameInteractions.java
@@ -1,7 +1,7 @@
 /*
 ASCII Art code is referenced from: https://www.baeldung.com/ascii-art-in-java
  */
-package game_state;
+package com.islandescape.controllers;
 
 import java.awt.Font;
 import java.awt.Graphics;
@@ -11,7 +11,7 @@ import java.awt.image.BufferedImage;
 import java.util.Scanner;
 import java.util.concurrent.TimeUnit;
 
-public class GameState {
+public class GameInteractions {
     private static final String gameTitle = "I" + " " + "S" + " " + "L" + " " + "A" + " " + "N" + " " + "D" +
             " " + " " + "E" + " " + "S" + " " + "C" + " " + "A" + " " + "P" + " " + "E"  ;
     private static final int width = 200;
@@ -73,7 +73,7 @@ public class GameState {
         char[] array = introductionText.toCharArray();
         for (char text : array) {
             System.out.print(text);
-            TimeUnit.MILLISECONDS.sleep(0);
+            TimeUnit.MILLISECONDS.sleep(30);
         }
     }
