+++ src/main/java/com/islandescape/entities/Item.java
@@ -1,4 +1,4 @@
-package entities;
+package com.islandescape.entities;
 
 import java.util.Objects;
