+++ src/main/java/com/islandescape/entities/MagicTotem.java
@@ -1,4 +1,4 @@
-package entities;
+package com.islandescape.entities;
 
 import java.io.BufferedReader;
 import java.io.IOException;
